[CC] Clean comp_parameters
